+++ src/main/java/com/example/adminshift/repository/EventRepository.java
@@ -9,5 +9,5 @@ import com.example.adminshift.entity.Event;
  * イベント情報のデータアクセスを提供するリポジトリ
  */
 @Repository
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
 }
